refactor: add provider folder, minor reorganization
