+++ frontend/src/lib/next.ts
@@ -1,0 +1,6 @@
+import { NextPage } from "next";
+import { ReactElement, ReactNode } from "react";
+
+export type NextPageWithLayout<P = Record<string, unknown>> = NextPage<P> & {
+  getLayout?: (page: ReactElement) => ReactNode;
+};
