+++ frontend/src/pages/about/itv.tsx
@@ -5,7 +5,7 @@ import { ItvCard } from "@/components/pages/about/ItvCard";
 import { ItvMember } from "@/components/pages/about/ItvCard/types";
 import { Template } from "@/components/pages/about/Template";
 import { Layout } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 const ItvPage: NextPageWithLayout = () => {
   const itvMembers: ItvMember[] = [
