+++ frontend/src/pages/about/organization.tsx
@@ -10,7 +10,7 @@ import { Link } from "@/components";
 import { TabPanel } from "@/components/pages/about/TabPanel";
 import { Template } from "@/components/pages/about/Template";
 import { Layout } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 import { getSortedPosts } from "@/utils/posts";
 
 type Post = {
