+++ frontend/src/pages/baksida/index.tsx
@@ -18,7 +18,7 @@ import { useRef } from "react";
 
 import { ContactInfo } from "@/components/pages/baksida/ContactInfo";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 import adrian from "~/public/img/adrian.jpg";
 import Amund from "~/public/img/Amund.jpg";
 import Laila from "~/public/img/Laila.jpg";
