+++ frontend/src/pages/cabins/index.tsx
@@ -10,7 +10,7 @@ import { FAQ } from "@/components/pages/cabins/Documents/FAQ";
 import { outsideImages } from "@/components/pages/cabins/ImageSlider/imageData";
 import { ImageSlider } from "@/components/pages/cabins/ImageSlider/ImageSlider";
 import { Layout } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 /** Front page for cabins. Includes info about the cabins and link to the booking page (cabins/book). */
 const CabinsPage: NextPageWithLayout = () => {
