+++ frontend/src/pages/events/new.tsx
@@ -5,7 +5,7 @@ import { Link } from "@/components";
 import { CreateEvent } from "@/components/pages/events/Create";
 import { UserOrganizationsDocument } from "@/generated/graphql";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 /** Component for showing the create event page. */
 const NewEventPage: NextPageWithLayout = () => {
