+++ frontend/src/pages/index.tsx
@@ -3,8 +3,7 @@ import { NoSsr } from "@mui/material";
 import { LandingHero, LandingListings, LandingPromo } from "@/components/landing";
 import { LandingSection } from "@/components/landing/LandingSection";
 import { Layout } from "@/layouts/Layout";
-
-import { NextPageWithLayout } from "./_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 const IndexPage: NextPageWithLayout = () => {
   return (
