+++ frontend/src/pages/listings/index.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 import { Listings } from "@/components/pages/listings/index/Listings";
 import { Title } from "@/components/Title/";
 import { Layout } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 // page to show all listings open to users
 const ListingsPage: NextPageWithLayout = () => {
