+++ frontend/src/pages/listings/new.tsx
@@ -1,6 +1,6 @@
 import { NewListing } from "@/components/pages/listings/organization/NewListing";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 /** Page for creating new listings, navigates to the newly created listing upon completion. */
 const NewListingPage: NextPageWithLayout = () => <NewListing />;
