+++ frontend/src/pages/orgs/[orgId]/listings/[listingId]/edit.tsx
@@ -9,7 +9,7 @@ import { useState } from "react";
 import { ListingForm } from "@/components/pages/listings/organization/ListingForm";
 import { ListingDocument, UpdateListingDocument } from "@/generated/graphql";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 import { ListingInput } from "@/types/listings";
 
 dayjs.extend(utc);
