+++ frontend/src/pages/orgs/[orgId]/listings/new.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 
 import { NewListing } from "@/components/pages/listings/organization/NewListing";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 /** Page for creating new listings, navigates to the newly created listing upon completion. */
 const NewListingPage: NextPageWithLayout = () => {
