+++ frontend/src/pages/profile/index.tsx
@@ -18,7 +18,7 @@ import {
 import { UserDocument } from "@/generated/graphql";
 import { Layout, RootStyle } from "@/layouts/Layout";
 import { addApolloState, initializeApollo } from "@/lib/apolloClient";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 import { generateFeideLoginUrl } from "@/utils/auth";
 
 const ID_PREFIX = "profile-";
