+++ frontend/src/pages/register/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 
 import { UserForm } from "@/components/pages/profile/UserForm";
 import { Layout, RootStyle } from "@/layouts/Layout";
-import { NextPageWithLayout } from "@/pages/_app";
+import { NextPageWithLayout } from "@/lib/next";
 
 const ID_PREFIX = "registerUser-";
