+++ frontend/src/providers/index.ts
@@ -1,0 +1,1 @@
+export { AppProvider } from "./AppProvider";
